# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for your invoice app.",
        "title": "Smart Invoice Generator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/protected/invoices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all invoices for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "List invoices by user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new invoice for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Create a new invoice",
                "parameters": [
                    {
                        "description": "Invoice data",
                        "name": "invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    }
                }
            }
        },
        "/v1/protected/invoices/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves an invoice by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Get invoice by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an invoice by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Update an invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invoice data",
                        "name": "invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/public/invoices/generate-pdf": {
            "post": {
                "description": "Generates and downloads the PDF for a given invoice ID",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Download invoice PDF",
                "parameters": [
                    {
                        "description": "Invoice data",
                        "name": "invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GeneratePublicInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    }
                }
            }
        },
        "/v1/public/invoices/{id}/pdf": {
            "get": {
                "description": "Generates and downloads the PDF for a given invoice ID",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Download invoice PDF",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.GenericResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateInvoiceRequest": {
            "type": "object",
            "required": [
                "client_id",
                "currency",
                "due_date",
                "invoice_number",
                "items",
                "tax_rate"
            ],
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR",
                        "IDR"
                    ]
                },
                "due_date": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InvoiceItemRequest"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "tax_rate": {
                    "type": "number"
                }
            }
        },
        "dto.GeneratePublicInvoiceRequest": {
            "type": "object",
            "required": [
                "currency",
                "due_date",
                "invoice_number",
                "issue_date",
                "recipient",
                "sender"
            ],
            "properties": {
                "currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR",
                        "IDR"
                    ]
                },
                "due_date": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "issue_date": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InvoiceItemUpdateRequest"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "recipient": {
                    "$ref": "#/definitions/dto.SenderRecipientRequest"
                },
                "sender": {
                    "$ref": "#/definitions/dto.SenderRequest"
                },
                "tax_rate": {
                    "type": "number"
                }
            }
        },
        "dto.InvoiceItemRequest": {
            "type": "object",
            "required": [
                "description",
                "quantity",
                "unit_price"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "unit_price": {
                    "type": "number"
                }
            }
        },
        "dto.InvoiceItemUpdateRequest": {
            "type": "object",
            "required": [
                "description",
                "quantity",
                "unit_price"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "unit_price": {
                    "type": "number"
                }
            }
        },
        "dto.SenderRecipientRequest": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "dto.SenderRequest": {
            "type": "object",
            "required": [
                "address",
                "bank_account_name",
                "bank_account_number",
                "bank_name",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "bank_account_name": {
                    "type": "string"
                },
                "bank_account_number": {
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateInvoiceRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InvoiceItemUpdateRequest"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tax_rate": {
                    "type": "number"
                }
            }
        },
        "utils.GenericResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}